+++ lib/libft/ft_strlen.c
@@ -1,7 +1,8 @@
+# include "lemin.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i] != '\0')
